+++ tests/integ/modin/binary/test_timedelta.py
@@ -11,8 +11,10 @@ import pandas as native_pd
 import pytest
 
 import snowflake.snowpark.modin.plugin  # noqa: F401
+from snowflake.snowpark.exceptions import SnowparkSQLException
 from tests.integ.modin.sql_counter import sql_count_checker
 from tests.integ.modin.utils import (
+    assert_series_equal,
     assert_snowpark_pandas_equals_to_pandas_without_dtypecheck,
     create_test_dfs,
     create_test_series,
@@ -112,9 +114,34 @@ def timestamp_scalar(request):
 @pytest.fixture(
     params=[
         pd.Timedelta("10 days 23:59:59.123456789"),
+        datetime.timedelta(microseconds=1),
+        datetime.timedelta(microseconds=2),
+        pd.Timedelta(nanoseconds=1),
+        pd.Timedelta(nanoseconds=2),
+        pd.Timedelta(nanoseconds=3),
+        pd.Timedelta(days=1),
+        pd.Timedelta(days=1, hours=1),
+        pd.Timedelta(days=10),
+    ]
+)
+def timedelta_scalar_positive(request):
+    return request.param
+
+
+@pytest.fixture(
+    params=[
+        pd.Timedelta("10 days 23:59:59.123456789"),
         pd.Timedelta("-10 days 23:59:59.123456789"),
         datetime.timedelta(days=-10, hours=23),
         datetime.timedelta(microseconds=1),
+        datetime.timedelta(microseconds=2),
+        pd.Timedelta(nanoseconds=1),
+        pd.Timedelta(nanoseconds=2),
+        pd.Timedelta(nanoseconds=3),
+        pd.Timedelta(days=1),
+        pd.Timedelta(days=1, hours=1),
+        pd.Timedelta(days=10),
+        pd.Timedelta(days=-1),
     ]
 )
 def timedelta_scalar(request):
@@ -135,6 +162,36 @@ def timedelta_dataframes_1() -> tuple[pd.DataFrame, native_pd.DataFrame]:
 
 
 @pytest.fixture
+def timedelta_dataframes_postive_no_nulls_1_2x2() -> tuple[
+    pd.DataFrame, native_pd.DataFrame
+]:
+    return create_test_dfs(
+        [
+            [pd.Timedelta(days=1), pd.Timedelta(days=4)],
+            [
+                pd.Timedelta(days=2),
+                pd.Timedelta(days=3),
+            ],
+        ]
+    )
+
+
+@pytest.fixture
+def timedelta_dataframes_with_negatives_no_nulls_1_2x2() -> tuple[
+    pd.DataFrame, native_pd.DataFrame
+]:
+    return create_test_dfs(
+        [
+            [pd.Timedelta(days=1), pd.Timedelta(days=4)],
+            [
+                pd.Timedelta(days=2),
+                pd.Timedelta(days=-3),
+            ],
+        ]
+    )
+
+
+@pytest.fixture
 def timedelta_series_1() -> tuple[pd.Series, native_pd.Series]:
     return create_test_series(
         [
@@ -148,10 +205,332 @@ def timedelta_series_1() -> tuple[pd.Series, native_pd.Series]:
     )
 
 
+@pytest.fixture
+def timedelta_series_positive_no_nulls_1_length_6() -> tuple[
+    pd.Series, native_pd.Series
+]:
+    return create_test_series(
+        [
+            pd.Timedelta(days=1),
+            pd.Timedelta(days=2),
+            pd.Timedelta(days=3),
+            pd.Timedelta(days=4),
+            pd.Timedelta(days=5),
+            pd.Timedelta(days=6),
+        ]
+    )
+
+
+@pytest.fixture
+def timedelta_series_no_nulls_2_length_6() -> tuple[pd.Series, native_pd.Series]:
+    return create_test_series(
+        [
+            pd.Timedelta(microseconds=7),
+            pd.Timedelta(hours=6, minutes=5),
+            pd.Timedelta(hours=4, minutes=3),
+            pd.Timedelta(hours=2, minutes=1),
+            pd.Timedelta(hours=8, minutes=9),
+            pd.Timedelta(hours=9, minutes=8),
+        ]
+    )
+
+
+@pytest.fixture
+def timedelta_series_no_nulls_3_length_2() -> tuple[pd.Series, native_pd.Series]:
+    return create_test_series(
+        [
+            pd.Timedelta(microseconds=7),
+            pd.Timedelta(hours=6, minutes=5),
+        ]
+    )
+
+
+@pytest.fixture(
+    params=[
+        "sub",
+        "rsub",
+        "add",
+        "radd",
+        "div",
+        "rdiv",
+        "truediv",
+        "rtruediv",
+        "floordiv",
+        "rfloordiv",
+        "mod",
+        "rmod",
+        "eq",
+        "ne",
+        "gt",
+        "lt",
+        "ge",
+        "le",
+    ]
+)
+def op_between_timedeltas(request) -> list[str]:
+    """Valid operations between timedeltas."""
+    return request.param
+
+
+class TestInvalid:
+    """
+    Test invalid binary operations, e.g. subtracting a timestamp from a timedelta.
+
+    For simplicity, check these cases for operations between dataframes and
+    scalars only.
+    """
+
+    @sql_count_checker(query_count=0)
+    def test_timedelta_scalar_minus_timestamp_dataframe(self):
+        eval_snowpark_pandas_result(
+            *create_test_dfs([datetime.datetime(year=2024, month=8, day=21)]),
+            lambda df: pd.Timedelta(1) - df,
+            expect_exception=True,
+            expect_exception_type=TypeError,
+            expect_exception_match=re.escape(
+                "bad operand type for unary -: 'DatetimeArray"
+            ),
+        )
+
+    @sql_count_checker(query_count=0)
+    @pytest.mark.parametrize(
+        "operation,error_symbol",
+        [("__or__", "|"), ("__ror__", "|"), ("__and__", "&"), ("__rand__", "&")],
+    )
+    def test_timedelta_dataframe_bitwise_operation_with_timedelta_scalar(
+        self, operation, timedelta_dataframes_1, error_symbol
+    ):
+        eval_snowpark_pandas_result(
+            *timedelta_dataframes_1,
+            lambda df: getattr(df, operation)(pd.Timedelta(2)),
+            expect_exception=True,
+            # pandas exception depends on the input types and is something like
+            # "unsupported operand type(s) for &: 'Timedelta' and 'TimedeltaArray'",
+            # but Snowpwark pandas always gives the same exception.
+            assert_exception_equal=False,
+            expect_exception_type=TypeError,
+            expect_exception_match=re.escape(
+                f"unsupported operand type for {error_symbol}: Timedelta"
+            ),
+        )
+
+    @sql_count_checker(query_count=0)
+    @pytest.mark.parametrize("operation", ["pow", "rpow"])
+    def test_timedelta_dataframe_exponentiation_with_timedelta_scalar(
+        self, operation, timedelta_dataframes_1
+    ):
+        eval_snowpark_pandas_result(
+            *timedelta_dataframes_1,
+            lambda df: getattr(df, operation)(pd.Timedelta(2)),
+            expect_exception=True,
+            # pandas exception depends on the input types and is something
+            # like "cannot perform __rpow__ with this index type:
+            # TimedeltaArray", but Snowpwark pandas always gives the same
+            # exception.
+            assert_exception_equal=False,
+            expect_exception_type=TypeError,
+            expect_exception_match=re.escape(
+                "unsupported operand type for **: Timedelta"
+            ),
+        )
+
+    @sql_count_checker(query_count=0)
+    @pytest.mark.parametrize("operation", ["mul", "rmul"])
+    def test_timedelta_dataframe_multiplied_by_timedelta_scalar_invalid(
+        self, operation, timedelta_dataframes_1
+    ):
+        eval_snowpark_pandas_result(
+            *timedelta_dataframes_1,
+            lambda df: getattr(df, operation)(pd.Timedelta(2)),
+            expect_exception=True,
+            expect_exception_type=np.core._exceptions._UFuncBinaryResolutionError,
+            expect_exception_match=re.escape(
+                "ufunc 'multiply' cannot use operands with types dtype('<m8[ns]') and dtype('<m8[ns]')"
+            ),
+        )
+
+
+class TestNumericEdgeCases:
+    """
+    Test numeric edge cases, e.g. arithmetic with zero and mod by a negative number.
+    """
+
+    @sql_count_checker(query_count=1)
+    @pytest.mark.parametrize(
+        "lhs,rhs,expected_pandas,expected_snow",
+        [(3, -2, -1, 1), (-3, -2, -1, -1), (-3, 2, 1, -1)],
+    )
+    def test_timedelta_mod_with_negative(
+        self, lhs, rhs, expected_pandas, expected_snow
+    ):
+        """Snowflake sometimes has different behavior for mod with negative numbers."""
+        snow_series, pandas_series = create_test_series(pd.Timedelta(lhs))
+        assert_series_equal(
+            pandas_series % pd.Timedelta(rhs),
+            native_pd.Series(pd.Timedelta(expected_pandas)),
+        )
+        assert_snowpark_pandas_equals_to_pandas_without_dtypecheck(
+            snow_series % pd.Timedelta(rhs),
+            native_pd.Series(pd.Timedelta(expected_snow)),
+        )
+
+    @sql_count_checker(query_count=0)
+    def test_divide_timedelta_by_zero_timedelta(self):
+        snow_series, pandas_series = create_test_series(pd.Timedelta(1))
+        assert_series_equal(pandas_series / pd.Timedelta(0), native_pd.Series(np.inf))
+        with pytest.raises(SnowparkSQLException, match=re.escape("Division by zero")):
+            (snow_series / pd.Timedelta(0)).to_pandas()
+
+    @sql_count_checker(query_count=0)
+    def test_floordiv_timedelta_by_zero_timedelta(self):
+        snow_series, pandas_series = create_test_series(pd.Timedelta(1))
+        assert_series_equal(pandas_series // pd.Timedelta(0), native_pd.Series(0))
+        with pytest.raises(SnowparkSQLException, match=re.escape("Division by zero")):
+            (snow_series // pd.Timedelta(0)).to_pandas()
+
+    @sql_count_checker(query_count=1)
+    def test_mod_timedelta_by_zero_timedelta(self):
+        snow_series, pandas_series = create_test_series(pd.Timedelta(1))
+        assert_series_equal(
+            pandas_series % pd.Timedelta(0),
+            native_pd.Series(1, dtype="timedelta64[ns]"),
+        )
+        assert_snowpark_pandas_equals_to_pandas_without_dtypecheck(
+            snow_series % pd.Timedelta(0),
+            native_pd.Series(pd.NaT, dtype="timedelta64[ns]"),
+        )
+
+    @sql_count_checker(query_count=1)
+    @pytest.mark.parametrize("op", ["add", "radd", "sub", "rsub"])
+    def test_timedelta_plus_or_minus_zero_timedelta(self, timedelta_dataframes_1, op):
+        eval_snowpark_pandas_result(
+            *timedelta_dataframes_1, lambda df: getattr(df, op)(pd.Timedelta(0))
+        )
+
+
+class TestNullsInTimedeltaComparisons:
+    """
+    Test comparisons (e.g. <=) between timedeltas where some operands are null.
+
+    Snowpark pandas returns null for comparsions where either operand is null,
+    whereas pandas considers null values to be equal to each other, and not
+    less than, greater than, or equal to any other values.
+
+    We test these cases separately because we can't use pandas to validate
+    Snowpark pandas results.
+    """
+
+    @sql_count_checker(query_count=1)
+    @pytest.mark.parametrize("op", ["eq", "ne", "lt", "le", "gt", "ge"])
+    def test_comparison_with_null(self, timedelta_dataframes_1, op):
+        snow_df = timedelta_dataframes_1[0]
+        assert_snowpark_pandas_equals_to_pandas_without_dtypecheck(
+            getattr(snow_df, op)(pd.NaT),
+            native_pd.DataFrame([[None, None], [None, None]]),
+        )
+
+    @sql_count_checker(query_count=1)
+    def test_eq_non_null(self, timedelta_dataframes_1):
+        snow_df = timedelta_dataframes_1[0]
+        assert_snowpark_pandas_equals_to_pandas_without_dtypecheck(
+            snow_df == pd.Timedelta(days=1),
+            native_pd.DataFrame([[True, None], [False, False]]),
+        )
+
+    @sql_count_checker(query_count=1)
+    def test_neq_non_null(self, timedelta_dataframes_1):
+        snow_df = timedelta_dataframes_1[0]
+        assert_snowpark_pandas_equals_to_pandas_without_dtypecheck(
+            snow_df != pd.Timedelta(days=1),
+            native_pd.DataFrame([[False, None], [True, True]]),
+        )
+
+
+@sql_count_checker(query_count=1)
+@pytest.mark.parametrize(
+    "op",
+    [
+        "sub",
+        "rsub",
+        "add",
+        "radd",
+    ],
+)
+def test_timedelta_timedelta_null_arithmetic_valid_in_pandas(
+    op, timedelta_dataframes_1
+):
+    """These timedelta arithmetic operations with null are valid in Snowpark pandas and pandas."""
+    eval_snowpark_pandas_result(
+        *timedelta_dataframes_1, lambda df: getattr(df, op)(pd.NaT)
+    )
+
+
+@sql_count_checker(query_count=1)
+@pytest.mark.parametrize("op", ["div", "rdiv", "mod", "rmod", "floordiv", "rfloordiv"])
+def test_timedelta_timedelta_null_arithmetic_invalid_in_pandas(
+    op, timedelta_dataframes_1
+):
+    """It is not valid in pandas to divide (div, rdiv, mod, rmod, floordiv, rfloordiv) timedelta with null scalar."""
+    snow_df, pandas_df = timedelta_dataframes_1
+
+    def op_with_null(df):
+        return getattr(df, op)(pd.NaT)
+
+    with pytest.raises(TypeError):
+        op_with_null(pandas_df)
+    assert_snowpark_pandas_equals_to_pandas_without_dtypecheck(
+        op_with_null(snow_df), native_pd.DataFrame([[pd.NaT, pd.NaT], [pd.NaT, pd.NaT]])
+    )
+
+
+@sql_count_checker(query_count=1)
+@sql_count_checker(query_count=1)
+@pytest.mark.parametrize(
+    "op",
+    [
+        "sub",
+        "rsub",
+        "add",
+        "radd",
+        "div",
+        "rdiv",
+        "truediv",
+        "rtruediv",
+        "floordiv",
+        "rfloordiv",
+    ],
+)
+def test_valid_arithmetic_between_possibly_negative_timedeltas(
+    timedelta_dataframes_with_negatives_no_nulls_1_2x2, op, timedelta_scalar
+):
+    """
+    Test that all valid arithmetic (except mod) between Timedeltas works with negative operands.
+
+    Test just dataframe left operand and scalar right operand in this test case.
+    """
+    if op == "rfloordiv" and timedelta_scalar in (
+        datetime.timedelta(microseconds=1),
+        datetime.timedelta(microseconds=2),
+        pd.Timedelta(nanoseconds=1),
+        pd.Timedelta(nanoseconds=2),
+        pd.Timedelta(nanoseconds=3),
+    ):
+        # snowflake returns 0 for dividing tiny timedeltas by large (~days)
+        # timedelta dataframes, so floordiv gives 0 in Snowflake but -1 in
+        # pandas. e.g.  pd.Timedelta('0 days 00:00:00.000000003') / pd.DataFrame([pd.Timedelta('-3 days')])
+        # translates to  pd.session.sql('SELECT 3.0 / -259200000000000.0').to_pandas(),
+        # which gives 0
+        pytest.xfail("precision difference in snowflake")
+    eval_snowpark_pandas_result(
+        *timedelta_dataframes_with_negatives_no_nulls_1_2x2,
+        lambda df: getattr(df, op)(timedelta_scalar),
+    )
+
+
 class TestDataFrameAndScalar:
     @pytest.mark.parametrize("operation", ["sub", "rsub"])
     @sql_count_checker(query_count=1)
-    def test_timestamp_col_minus_timestamp_scalar(
+    def test_timestamp_minus_timestamp(
         self,
         timestamp_scalar,
         operation,
@@ -194,18 +573,6 @@ class TestDataFrameAndScalar:
             lambda df: getattr(df, operation)(timedelta_scalar),
         )
 
-    @sql_count_checker(query_count=0)
-    def test_timedelta_scalar_minus_timestamp_dataframe_negative(self):
-        eval_snowpark_pandas_result(
-            *create_test_dfs([datetime.datetime(year=2024, month=8, day=21)]),
-            lambda df: pd.Timedelta(1) - df,
-            expect_exception=True,
-            expect_exception_type=TypeError,
-            expect_exception_match=re.escape(
-                "bad operand type for unary -: 'DatetimeArray"
-            ),
-        )
-
     @sql_count_checker(query_count=1)
     @pytest.mark.parametrize("operation", ["add", "radd"])
     def test_timedelta_dataframe_plus_timestamp_scalar(
@@ -215,13 +582,30 @@ class TestDataFrameAndScalar:
             *timedelta_dataframes_1, lambda df: getattr(df, operation)(timestamp_scalar)
         )
 
-    @pytest.mark.parametrize("operation", ["sub", "rsub", "add", "radd"])
-    @pytest.mark.xfail(strict=True, raises=NotImplementedError, reason="SNOW-1637101")
-    def test_timedelta_dataframe_plus_or_minus_timedelta_scalar(
-        self, timedelta_dataframes_1, timedelta_scalar, operation
+    @sql_count_checker(query_count=1)
+    def test_timedelta_dataframe_with_timedelta_scalar(
+        self,
+        timedelta_dataframes_postive_no_nulls_1_2x2,
+        timedelta_scalar_positive,
+        op_between_timedeltas,
     ):
         eval_snowpark_pandas_result(
-            *timedelta_dataframes_1, lambda df: getattr(df, operation)(timedelta_scalar)
+            *timedelta_dataframes_postive_no_nulls_1_2x2,
+            lambda df: getattr(df, op_between_timedeltas)(timedelta_scalar_positive),
+        )
+
+    @pytest.mark.xfail(
+        strict=True,
+        raises=NotImplementedError,
+        match="does not yet support the binary operation .* with a Timedelta column and a non-Timedelta column",
+        reason="SNOW-1637102",
+    )
+    @pytest.mark.parametrize("op", ["mul", "rmul"])
+    def test_timedelta_dataframe_with_integer_scalar(
+        self, timedelta_dataframes_postive_no_nulls_1_2x2, op
+    ):
+        eval_snowpark_pandas_result(
+            *timedelta_dataframes_postive_no_nulls_1_2x2, lambda df: getattr(df, op)(3)
         )
 
 
@@ -266,6 +650,20 @@ class TestSeriesAndScalar:
         eval_snowpark_pandas_result(
             *timedelta_series_1,
             lambda series: getattr(series, operation)(timestamp_scalar),
+        )
+
+    @sql_count_checker(query_count=1)
+    def test_timedelta_series_with_timedelta_scalar(
+        self,
+        timedelta_scalar_positive,
+        timedelta_series_positive_no_nulls_1_length_6,
+        op_between_timedeltas,
+    ):
+        eval_snowpark_pandas_result(
+            *timedelta_series_positive_no_nulls_1_length_6,
+            lambda series: getattr(series, op_between_timedeltas)(
+                timedelta_scalar_positive
+            ),
         )
 
 
@@ -313,6 +711,20 @@ class TestDataFrameAndListLikeAxis1:
                 [
                     pd.Timestamp(1, unit="D"),
                     pd.Timestamp(-13, unit="D"),
+                ]
+            ),
+        )
+
+    @sql_count_checker(query_count=1)
+    def test_timedelta_dataframe_with_timedelta_list_like(
+        self, op_between_timedeltas, timedelta_dataframes_postive_no_nulls_1_2x2
+    ):
+        eval_snowpark_pandas_result(
+            *timedelta_dataframes_postive_no_nulls_1_2x2,
+            lambda df: getattr(df, op_between_timedeltas)(
+                [
+                    pd.Timedelta(days=2, milliseconds=140),
+                    pd.Timedelta(days=1, milliseconds=770),
                 ]
             ),
         )
@@ -368,6 +780,24 @@ class TestSeriesAndListLike:
             ),
         )
 
+    @sql_count_checker(query_count=1, join_count=1)
+    def test_timedelta_series_and_timedelta_list_like(
+        self, op_between_timedeltas, timedelta_series_positive_no_nulls_1_length_6
+    ):
+        eval_snowpark_pandas_result(
+            *timedelta_series_positive_no_nulls_1_length_6,
+            lambda series: getattr(series, op_between_timedeltas)(
+                [
+                    pd.Timedelta(days=6, minutes=6),
+                    pd.Timedelta(days=5, minutes=5),
+                    pd.Timedelta(days=4, minutes=4),
+                    pd.Timedelta(days=3, minutes=3),
+                    pd.Timedelta(days=2, minutes=2),
+                    pd.Timedelta(nanoseconds=2),
+                ]
+            ),
+        )
+
 
 class TestDataFrameAndListLikeAxis0:
     @sql_count_checker(query_count=1, join_count=1)
@@ -415,6 +845,21 @@ class TestDataFrameAndListLikeAxis0:
                 [
                     pd.Timestamp(1, unit="D"),
                     pd.Timestamp(999, unit="D"),
+                ],
+                axis=0,
+            ),
+        )
+
+    @sql_count_checker(query_count=1, join_count=1)
+    def test_timedelta_dataframe_and_timedelta_list_like(
+        self, op_between_timedeltas, timedelta_dataframes_postive_no_nulls_1_2x2
+    ):
+        eval_snowpark_pandas_result(
+            *timedelta_dataframes_postive_no_nulls_1_2x2,
+            lambda df: getattr(df, op_between_timedeltas)(
+                [
+                    pd.Timedelta(nanoseconds=1),
+                    pd.Timedelta(microseconds=2),
                 ],
                 axis=0,
             ),
@@ -514,6 +959,21 @@ class TestSeriesAndSeries:
             lambda inputs: getattr(inputs[0], op)(inputs[1]),
         )
 
+    @sql_count_checker(query_count=1, join_count=1)
+    def test_timedelta_and_timedleta(
+        self,
+        op_between_timedeltas,
+        timedelta_series_positive_no_nulls_1_length_6,
+        timedelta_series_no_nulls_2_length_6,
+    ):
+        snow_lhs, pandas_lhs = timedelta_series_positive_no_nulls_1_length_6
+        snow_rhs, pandas_rhs = timedelta_series_no_nulls_2_length_6
+        eval_snowpark_pandas_result(
+            (snow_lhs, snow_rhs),
+            (pandas_lhs, pandas_rhs),
+            lambda inputs: getattr(inputs[0], op_between_timedeltas)(inputs[1]),
+        )
+
 
 class TestDataFrameAndSeriesAxis0:
     @pytest.mark.parametrize("op", ["sub", "rsub"])
@@ -577,6 +1037,21 @@ class TestDataFrameAndSeriesAxis0:
             (snow_df, snow_series),
             (pandas_df, pandas_series),
             lambda t: getattr(t[0], op)(t[1], axis=0),
+        )
+
+    @sql_count_checker(query_count=1, join_count=1)
+    def test_timedelta_dataframe_and_timedelta_series(
+        self,
+        op_between_timedeltas,
+        timedelta_dataframes_postive_no_nulls_1_2x2,
+        timedelta_series_no_nulls_3_length_2,
+    ):
+        snow_df, pandas_df = timedelta_dataframes_postive_no_nulls_1_2x2
+        snow_series, pandas_series = timedelta_series_no_nulls_3_length_2
+        eval_snowpark_pandas_result(
+            (snow_df, snow_series),
+            (pandas_df, pandas_series),
+            lambda t: getattr(t[0], op_between_timedeltas)(t[1], axis=0),
         )
 
 
@@ -669,7 +1144,7 @@ class TestDataFrameAndSeriesAxis1:
         )
 
     @sql_count_checker(
-        # One query to materialize the series for the subtraction, and another
+        # One query to materialize the series for the operation, and another
         # query to materialize the result.
         query_count=2
     )
@@ -767,7 +1242,7 @@ class TestDataFrameAndSeriesAxis1:
         )
 
     @sql_count_checker(
-        # One query to materialize the series for the subtraction, and another
+        # One query to materialize the series for the operation, and another
         # query to materialize the result.
         query_count=2
     )
@@ -809,6 +1284,25 @@ class TestDataFrameAndSeriesAxis1:
                     ],
                 ]
             ),
+        )
+
+    @sql_count_checker(
+        # One query to materialize the series for the subtraction, and another
+        # query to materialize the result.
+        query_count=2
+    )
+    def test_timedelta_dataframe_with_timedelta_series(
+        self,
+        timedelta_dataframes_postive_no_nulls_1_2x2,
+        timedelta_series_no_nulls_3_length_2,
+        op_between_timedeltas,
+    ):
+        snow_df, pandas_df = timedelta_dataframes_postive_no_nulls_1_2x2
+        snow_series, pandas_series = timedelta_series_no_nulls_3_length_2
+        eval_snowpark_pandas_result(
+            (snow_df, snow_series),
+            (pandas_df, pandas_series),
+            lambda t: getattr(t[0], op_between_timedeltas)(t[1]),
         )
 
 
@@ -890,4 +1384,24 @@ class TestDataFrameAndDataFrameAxis1:
             (snow_lhs, snow_rhs),
             (pandas_lhs, pandas_rhs),
             lambda t: getattr(t[0], op)(t[1]),
+        )
+
+    @sql_count_checker(query_count=1, join_count=1)
+    def test_two_timedelta_datfaframes(
+        self, timedelta_dataframes_postive_no_nulls_1_2x2, op_between_timedeltas
+    ):
+        snow_lhs, pandas_lhs = timedelta_dataframes_postive_no_nulls_1_2x2
+        snow_rhs, pandas_rhs = create_test_dfs(
+            [
+                [
+                    pd.Timedelta(nanoseconds=1),
+                    pd.Timedelta(microseconds=1),
+                ],
+                [pd.Timedelta(days=2), pd.Timedelta(days=1)],
+            ]
+        )
+        eval_snowpark_pandas_result(
+            (snow_lhs, snow_rhs),
+            (pandas_lhs, pandas_rhs),
+            lambda t: getattr(t[0], op_between_timedeltas)(t[1]),
         )